[COMMIT] - 어드민그룹관리 추가.

+++ src/main/java/com/example/cms/domain/admingroup/service/AdminGroupService.java
@@ -1,8 +1,11 @@
 package com.example.cms.domain.admingroup.service;
 
+import com.example.cms.domain.admin.repository.AdminRepository;
 import com.example.cms.domain.admingroup.dto.AdminGroupDTO;
 import com.example.cms.domain.admingroup.entity.AdminGroup;
 import com.example.cms.domain.admingroup.repository.AdminGroupRepository;
+import com.example.cms.system.util.MessageUtil;
+import com.example.cms.web.controller.group.AdminGroupForm;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,6 +21,8 @@ import java.util.Optional;
 public class AdminGroupService {
     
     private final AdminGroupRepository adminGroupRepository;
+    private final AdminRepository adminRepository;
+    private final MessageUtil messageUtil;
     
     @Transactional
     public Long save(AdminGroup adminGroup) {
@@ -30,6 +35,10 @@ public class AdminGroupService {
 
     @Transactional
     public void deleteById(Long adminGroupId) {
+        AdminGroup adminGroup = adminGroupRepository.findById(adminGroupId).orElseThrow(() -> new IllegalArgumentException(messageUtil.getMessage("message.admin.group.common.unknown.target")));
+        if (adminRepository.findByAdminGroup(adminGroup).isPresent()) {
+            throw new AdminGroupInUseException(messageUtil.getMessage("message.admin.group.in-use"));
+        }
         adminGroupRepository.deleteById(adminGroupId);
     }
 
@@ -39,5 +48,12 @@ public class AdminGroupService {
 
     public Page<AdminGroupDTO> findAdminGroupList(Pageable pageable) {
         return adminGroupRepository.findAdminGroupList(pageable);
+    }
+
+    @Transactional
+    public void update(AdminGroupForm adminGroupForm) {
+        AdminGroup adminGroup = adminGroupRepository.findById(Long.valueOf(adminGroupForm.getAdminGroupId()))
+                .orElseThrow(() -> new IllegalArgumentException(messageUtil.getMessage("message.admin.group.common.unknown.target")));
+        adminGroup.updateAdminGroup(adminGroupForm.getName(), adminGroupForm.getDescription());
     }
 }
